Better error messages for the mark5 readers
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@1540 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ src/mark5a_header.cc
@@ -199,7 +199,7 @@ bool Mark5a_header::checkCRC(uint8_t mask) {
   for (size_t i=0; i<12*N; i++) {
     if (crcBlock[i] != 0) {
       std::cout << "Error in crc " << i << " "
-      << std::hex << crcBlock[i] << std::dec
+      << std::hex << (int)crcBlock[i] << std::dec
       << std::endl;
       return false;
     }
